requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile requirements.in
#
amqp==5.0.9
    # via kombu
asgiref==3.5.0
    # via django
attrs==21.4.0
    # via pytest
billiard==3.6.4.0
    # via celery
black==22.1.0
    # via -r requirements.in
celery==5.2.3
    # via -r requirements.in
certifi==2021.10.8
    # via
    #   requests
    #   sentry-sdk
cfgv==3.3.1
    # via pre-commit
charset-normalizer==2.0.11
    # via requests
click==8.0.3
    # via
    #   black
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
deprecated==1.2.13
    # via redis
distlib==0.3.4
    # via virtualenv
django==4.0.1
    # via
    #   -r requirements.in
    #   django-cors-headers
    #   django-debug-toolbar
    #   django-extensions
    #   django-filter
    #   django-sql-utils
    #   djangorestframework
    #   djangorestframework-simplejwt
    #   drf-base64
    #   drf-nested-routers
    #   drf-yasg
django-cors-headers==3.11.0
    # via -r requirements.in
django-debug-toolbar==3.2.4
    # via -r requirements.in
django-extensions==3.1.5
    # via -r requirements.in
django-filter==21.1
    # via -r requirements.in
django-js-asset==1.2.2
    # via
    #   -r requirements.in
    #   django-mptt
django-mptt==0.13.4
    # via -r requirements.in
django-sql-utils==0.6.1
    # via -r requirements.in
djangorestframework==3.13.1
    # via
    #   -r requirements.in
    #   djangorestframework-simplejwt
    #   drf-base64
    #   drf-nested-routers
    #   drf-yasg
    #   pytest-drf
djangorestframework-simplejwt==5.0.0
    # via -r requirements.in
drf-base64==2.0
    # via -r requirements.in
drf-nested-routers==0.93.4
    # via -r requirements.in
drf-writable-nested==0.6.3
    # via -r requirements.in
drf-yasg==1.20.0
    # via -r requirements.in
filelock==3.4.2
    # via virtualenv
flake8==4.0.1
    # via -r requirements.in
gunicorn==20.1.0
    # via -r requirements.in
identify==2.4.8
    # via pre-commit
idna==3.3
    # via requests
inflection==0.3.1
    # via
    #   drf-yasg
    #   pytest-drf
iniconfig==1.1.1
    # via pytest
isort==5.10.1
    # via -r requirements.in
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
kombu==5.2.3
    # via celery
lazy-object-proxy==1.7.1
    # via pytest-common-subject
markupsafe==2.0.1
    # via jinja2
mccabe==0.6.1
    # via flake8
mypy-extensions==0.4.3
    # via black
nodeenv==1.6.0
    # via pre-commit
numpy==1.22.2
    # via pandas
packaging==21.3
    # via
    #   drf-yasg
    #   pytest
    #   redis
pandas==1.4.0
    # via -r requirements.in
pathspec==0.9.0
    # via black
pillow==9.0.0
    # via -r requirements.in
platformdirs==2.4.1
    # via
    #   black
    #   virtualenv
pluggy==1.0.0
    # via pytest
pre-commit==2.17.0
    # via -r requirements.in
prompt-toolkit==3.0.27
    # via click-repl
psycopg2-binary==2.9.3
    # via -r requirements.in
py==1.11.0
    # via pytest
pycodestyle==2.8.0
    # via flake8
pyflakes==2.4.0
    # via flake8
pygraphviz==1.9
    # via -r requirements.in
pyjwt==2.3.0
    # via djangorestframework-simplejwt
pyparsing==3.0.7
    # via packaging
pytest==6.2.5
    # via
    #   -r requirements.in
    #   pytest-common-subject
    #   pytest-django
    #   pytest-drf
    #   pytest-fixture-order
    #   pytest-lambda
pytest-assert-utils==0.3.0
    # via pytest-drf
pytest-common-subject==1.0.5
    # via pytest-drf
pytest-django==4.5.2
    # via -r requirements.in
pytest-drf==1.1.2
    # via -r requirements.in
pytest-fixture-order==0.1.3
    # via pytest-common-subject
pytest-lambda==1.2.5
    # via
    #   pytest-common-subject
    #   pytest-drf
python-dateutil==2.8.2
    # via pandas
python-environ==0.4.54
    # via -r requirements.in
pytz==2021.3
    # via
    #   celery
    #   djangorestframework
    #   pandas
pyyaml==6.0
    # via pre-commit
redis==4.1.2
    # via -r requirements.in
requests==2.27.1
    # via
    #   -r requirements.in
    #   coreapi
ruamel-yaml==0.17.20
    # via drf-yasg
ruamel-yaml-clib==0.2.6
    # via ruamel-yaml
sentry-sdk==1.5.4
    # via -r requirements.in
six==1.16.0
    # via
    #   click-repl
    #   python-dateutil
    #   virtualenv
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
    #   django-sql-utils
toml==0.10.2
    # via
    #   pre-commit
    #   pytest
tomli==2.0.1
    # via black
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.8
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
virtualenv==20.13.1
    # via pre-commit
wcwidth==0.2.5
    # via prompt-toolkit
wrapt==1.13.3
    # via
    #   deprecated
    #   pytest-lambda

# The following packages are considered to be unsafe in a requirements file:
# setuptools
